Update form controls `readonly` style.
Since `readonly` is non-editable, must be able to select and the data will be send when form submitted.
So, it should look like normal form controls but make it more noticable that it is read only by a little gray out the background color.

Summary: remove `not-allowed` cursor, lighter gray background.